feat(flow): enhance function reference handling and connection logic
- Added `fnRefsHash` to `FlowNode` type for efficient tracking of function references.
- Implemented memoization for connection states in `FlowNodeInner` to optimize rendering based on function reference connections.
- Introduced new helper functions in `flow-board-helpers.ts` to manage function reference connections and regular pin connections.
- Updated `handlePlaceNode` and `handlePlacePlaceholder` to support function reference connections when placing nodes.
- Enhanced `handleConnection` and `handleEdgesChange` to manage function reference edges correctly.
- Updated `parseBoard` to compute and store a hash of function references for nodes, ensuring accurate updates and rendering.
